removed isShaped and isPicturedisk
removed isShaped and isPicturedisk because I was unable properly validate the variables in the recordRegistrationRequest. what would happen is that if no value was given the booleans would be set to false. I could not check with my if else statement in the updateRecord method if the user had assigned them a vallue or not, sice the boolean could not return null, only false. So I chose to remove them, will add them back if I have a solution for this problem